Move TS outdir to `dist/`
