+++ .gitignore
@@ -15,4 +15,4 @@
 /node_modules/
 /typescript/
 /coverage/
-/lib/
+/dist/
