--- index.d.ts
@@ -1,6 +1,0 @@
-export { Application } from "./lib/application";
-export { CliApplication } from "./lib/cli";
-export { EventDispatcher, Event } from "./lib/utils/events";
-export { resetReflectionID } from "./lib/models/reflections/abstract";
-export { normalizePath } from "./lib/utils/fs";
-export { ProjectReflection } from "./lib/models/reflections/project";
