--- index.js
@@ -1,15 +1,0 @@
-"use strict";
-var application_1 = require("./lib/application");
-exports.Application = application_1.Application;
-var cli_1 = require("./lib/cli");
-exports.CliApplication = cli_1.CliApplication;
-var events_1 = require("./lib/utils/events");
-exports.EventDispatcher = events_1.EventDispatcher;
-exports.Event = events_1.Event;
-var abstract_1 = require("./lib/models/reflections/abstract");
-exports.resetReflectionID = abstract_1.resetReflectionID;
-var fs_1 = require("./lib/utils/fs");
-exports.normalizePath = fs_1.normalizePath;
-var project_1 = require("./lib/models/reflections/project");
-exports.ProjectReflection = project_1.ProjectReflection;
-//# sourceMappingURL=index.js.map
