+++ test/renderer.js
@@ -1,4 +1,4 @@
-var TypeDoc = require("../index.js");
+var TypeDoc = require("../");
 var FS      = require("fs-extra");
 var Path    = require("path");
 var Assert  = require("assert");
