+++ test/renderer/specs/index.html
@@ -80,13 +80,6 @@
 				<pre><code class="lang-bash">$ npm install typedoc --global
 $ typedoc
 </code></pre>
-				<h2 id="preview">Preview</h2>
-				<p>If you want to know what a documentation created with TypeDoc looks like, head over
-					to the homepage of the project. We&#39;ve setup examples demonstrating the two default
-				themes shipped with the package:</p>
-				<p><a href="http://typedoc.io/themes/default">http://typedoc.io/themes/default</a><br>
-				<a href="http://typedoc.io/themes/minimal">http://typedoc.io/themes/minimal</a></p>
-				<p>The default themes can be found here: <a href="https://github.com/TypeStrong/typedoc-default-themes">https://github.com/TypeStrong/typedoc-default-themes</a></p>
 				<h2 id="usage">Usage</h2>
 				<h3 id="shell">Shell</h3>
 				<p>TypeDoc accepts most of the command line arguments that the TypeScript compiler accepts. One major
@@ -184,7 +177,7 @@ $ typedoc
 					an issue or create a pull requests:<br>
 				<a href="https://github.com/TypeStrong/typedoc">https://github.com/TypeStrong/typedoc</a></p>
 				<h2 id="license">License</h2>
-				<p>Copyright (c) 2015 <a href="http://www.TypeDoc.de">Sebastian Lenz</a>.<br>
+				<p>Copyright (c) 2015 <a href="http://typedoc.org">Sebastian Lenz</a>.<br>
 				Licensed under the Apache License 2.0.</p>
 			</div>
 		</div>
