+++ test/typedoc.js
@@ -1,4 +1,4 @@
-var TypeDoc = require("../index.js");
+var TypeDoc = require("../");
 var Path    = require("path");
 var Assert  = require("assert");
