+++ tsconfig.json
@@ -1,23 +1,19 @@
 {
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "ES5",
-        "noImplicitAny": false,
-        "removeComments": true,
-        "preserveConstEnums": true,
-        "declaration": true,
-        "sourceMap": true,
-        "inlineSources": true, // required for debugging for npm installs
-        "outDir": "./",
-        "rootDir": "./src",
-        "experimentalDecorators": true
-    },
-    "include": [
-        "src/**/*.ts"
-    ],
-    "exclude": [
-    ],
-    "atom": {
-        "rewriteTsconfig": true
-    }
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "ES5",
+    "noImplicitAny": false,
+    "removeComments": true,
+    "preserveConstEnums": true,
+    "declaration": true,
+    "sourceMap": true,
+    "inlineSources": true, // Required for debugging NPM published package.
+    "outDir": "dist/",
+    "rootDir": "src/",
+    "experimentalDecorators": true
+  },
+  "include": [
+    "src/**/*.ts"
+  ],
+  "exclude": []
 }
